Simplified  and extracted to other file.
